Manually add all enum values to old apis.
Previous version of the apicheck tool that generated XML files did
not account for enum values. As such, DocLava cannot determine the API
level in which a given enum value was added or deprecated. Fortunately,
beginning with API level 14, this is resolved and the TXT files from
apicheck include enums. But to make the API level info visible in javadoc
for older APIs, I went back and added all the enum <field> values. Weeee!

Once this is safely submitted, I'll submit a small change to DocLava
that actually takes advantage of this info to show the "since" levels
in javadoc (because DocLava also had a bug in which it didn't create
the HDF data for enum values).

Change-Id: I1f88e8f9f3e2655d17eea628611318e207cbb535