Updated prebuilts for RenderScript compatibility library.
Bug: 10653601
Bug: 13751317

This change also updates the rules for using our new 64-bit prebuilts.
This is a squashed revert of 2 prior CLs (one for Linux, and another for
darwin). It also updates the rules to denote that these prebuilts are
for 64-bit use only.

Switch to LOCAL_MULTILIB := 64 for our 64-bit LLVM libraries.

Change-Id: I184d76641296f7539732529b265ad71ced9be747