Fix searching for overridden methods.
If the superclass is an abstract class, we want to search it for
possibly overridden methods; we don't want to search ourselves.

This bug resulted in (incorrectly) thinking that overridden "default"
methods from interfaces were standalone methods on the class.  (The
expected behavior has always been to omit any overridden methods from
the public API surface area to avoid redundant noise.)

Also fix another bug that was quietly letting overridden methods
change return types or exceptions without calling them out as API
changes.

This change is simply doing:

$ cd frameworks/base/ && git reset --hard goog/oc-dev
$ mmm -j32 external/doclava/ && make -j32 update-api
$ cp api/current.txt ../../prebuilts/sdk/api/27.txt
$ cp api/system-current.txt ../../prebuilts/sdk/system-api/27.txt

The current goog/oc-dev ToT as of this CL is
88f98985d5863cd4a3a1695648482be63c8c73bc.

Test: make update-api
Bug: 62675475
Change-Id: Iad17738804323a4cb544fa977a2c8dd035b2a4b9