searchKeyphrase* attributes are deleted from system API
The attributes are listed in public.xml *and* marked as @SystemApi in
attrs.xml. The former made it possible for clients to reference these
attributes via 'android:searchKeyphrase*' in their xml files and the
latter made it possible for clients to reference them via
android.R.attr.searchKeyphrase*. However, the former was the only use
case intended (and in fact the latter wasn't used at all.)

Since marking them as @SystemApi has been effectively no-op and will
even cause CTS failure when we do the exact match for SystemAPIs, the
markers are now removed.

Bug: 67891551
Test: m -j checkapi does not cause any problem
Change-Id: I66c23d2e76077362222e3c746405894ab75278f3