Mark prebuilt API stubs as java:sdk rather than java:platform
These prebuilt API stubs are used when building the unbundled apps
(TARGET_BUILD_APPS is not empty). However, the prebuilts have been
defined without LOCAL_SDK_VERSION, and as a result, their link-type has
been incorrectly identified as java:platform.

This is causing link-type check warning when building the unbundled
apps. And this warning is going to be promoted to error.

Fixing the error by adding LOCAL_SDK_VERSION := current or <ver> when
defining the prebuilts. Note that the same API stubs that are built from
the source code are already with LOCAL_SDK_VERSION set. This change
eliminates the inconsistency.

Bug: 69899800
Test: m -j ANDROID_BUILD_SPEC=vendor/google/build/app_build_spec.mk
Change-Id: I828eb1cfc1e14c7d72e2f08da7f68809348946bb