Update prebuilt script to import framework stubs into a new format.
The 'current', and finalized "N" dirs are moving to a new format, with the
subdirectories "core", "public", "system", where all the artifacts with a
particular api visibility level are stored. For the finalized dirs, the api
txt artifacts are also stored in the appropriate subdir.

For now, both formats will co-exist to allow migrating all the tools that
consume these files to be updated incrementally, before deleting the old files.

See this doc for more details:
https://docs.google.com/document/d/10vPR5VmwwbNJ-SugrQ7idtwfCBB7zImQ9aoedmHWj14/edit

Bug: 77525052
Test: ./update_prebuilts.py -p -f 99 4707605

Change-Id: Ib1de6f72c09c6aab3b1c0214d3fb35a83d4725c9