Add org.apache.http.legacy.jar stubs to current/system
org.apache.http.legacy is now built using java_sdk_library. As a result,
when a java module is built unbundled or built with a versioned SDK,
it is linked to the stubs lib in prebuilt/sdk.

Until now, we only had the stubs lib for org.apache.http.legacy in
current/public which is used only when LOCAL_SDK_VERSION is set to
current. For other cases (i.e. LOCAL_SDK_VERSION is unset or set to
system_current), we need the stubs in current/system directory as well.

Also, API txt files for the library is added for 28 (P), which is the
latest released API version as of now.

Bug: 77577799
Test: m -j

Merged-In: I1a3d9032c2223c519775d0a0bad57ca1d1996c41
Change-Id: I1a3d9032c2223c519775d0a0bad57ca1d1996c41
(cherry picked from commit 682e1f157b58235e9571b01a26c8d9706430a258)