Remove sdk_artifacts_dict
All api txt files and jar files are located in apistubs dir of Artifacts
(except android.jar). So we don't need to know the name of the sdk
library and we don't need sdk_artifacts_dict anymore.

sdk_artifacts_dict is removed, and then after all the files in apistubs
dir are received, then files of sdk_library for google only libraries
are removed.

And framework_sdk_target is changed to sdk_phone_armv7-sdk because
sdk_phone_armv7-sdk_mac was removed in go/ab

Bug: 77577799
Test: ./update_prebuilts.py -f 34 5019800
Change-Id: I70ddb191b5204b23a65ac6f4699caec4d6de274e