Regenerate historical signature files from android.jar files
The signature files in prebuilts/sdk/ were generated at the time
with the current version of doclava at the time. This means that
they suffer from bugs that have since been fixed in doclava, such
as missing type parameters, missing annotation member methods, and
so on. And prior to API level 14, the signature files were in an
XML format instead.

This CL regenerates all the signature files to reflect what was
actually in the android.jar files that we released.

This was done using metalava which reads in the API from the
actual released android.jar files instead of from sources. You
can reproduce the output by invoking metalava with these flags:

   --write-android-jar-signatures <android source dir> --format=v1

This CL writes the signature files back out (and for lower API levels
replaces XML files with text files) in the old, doclava-compatible
format. A followup CL will switch the format over to --format=v2
instead.

If you're studying the diffs, some things to notice:
- Recent versions of doclava started dropping the "final" modifier
  on members if the containing class is final
- Type parameters both for classes and methods were missing in
  older versions
- Older versions of the signature files included package private
  super classes with public members; these are now instead inlined
  into the API class (example: StringBuilder and
  AbstractStringBuilder).

Fixes: 116589402 Switch signature files over to the v2 format
Exempt-From-Owner-Approval: Large-scale tooling change
Test: make sdk
Change-Id: I81ec12166fcc319f1eda6fe367a4a4a73b756e59