Print the command being ran in fetch.
Error message being printed is regardless of the actual error.
Knowing the command that was run means the user can run it locally
and see the actual error for themselves.

Also updates target names from "support_library" to "androidx",
and the m2repository file name to the correct one as well.

Test: just run it.

Change-Id: I2916fb1dd5dd311f850823d209b099ce9b4e5706