Add missing java_sdk_library_imports for current/*.jar
The prebuilt versions of various libraries relied on logic in the
java_sdk_library module to make them available for unbundled builds.
That logic is being moved into java_sdk_library_import and so it
requires a java_sdk_library_import to be defined for each of the
libraries.

Note, not every jar that is present in the public/system/test
directory is a java_sdk_library.

Bug: 148080325
Test: m droid && TARGET_BUILD_APPS=Camera2 m
Change-Id: I14140e46248da058367201175eace69fcb670260