Add libraries to sdk, copied from clang-r365631b; add lld trampoline
lld depends on the libraries.

The darwin and linux trampoline allows lld to be invoked from the tool
root while residing in lld-bin, which is the correct location relative
to the libraries.

- darwin libc++.1.dylib, trampoline
- linux libc++.so.1, trampoline

This required temporarily reverting
I09219d4275b81c5aff3a8562abfd341300303d8e which had removed that version
of linux clang prebuilts.

Bug: 148267171
Bug: 142590626
Bug: 144040336
Test: N/A
Change-Id: Ied2992a01ed65953630466d6b1a6c425a441e3d1