Add resources and notices to 30/non-updatable.jar
The way the build worked When 30 was finalized, some parts of the
non-updatable framework was not part of the android-non-updatable stub
files, but were only merged into the final android.jar.

With SDK Extensions, we want to be able to create a new android.jar
which is a merge of the non-updatable framework from 30 but recent
module stubs. Making this work is made simpler if all the non-updatable
parts are in one place.

I ran these commands:

    sdk=$(readlink -f prebuilts/sdk/30)
    t=$(mktemp -d)
    unzip -qd $t $sdk/public/android.jar \
        NOTICES/* \
        assets/* \
        res/* \
        AndroidManifest.xml \
        resources.arsc
    cd $t
    zip -qur $sdk/public/android-non-updatable.jar *
    zip -qur $sdk/system/android-non-updatable.jar *
    zip -qur $sdk/module-lib/android-non-updatable.jar *

Bug: 184839225
Test: presubmit
Change-Id: I5ea8426a57cef14d7ff5dc84dd570b12cc25b6ea