Add --sdk_target option to update_prebuilts.py
Previously, when specifying -p or -f the build target was hard coded to
sdk_phone_armv7-win_sdk. Unfortunately, that target does not exist on
many branches making it necessary to temporarily modify the script to
change the target before updating the prebuilts. That meant the
information about the build target that was used was not part of the
commit message making it hard to track the origins of the prebuilts.

This change adds a --sdk_target option which allows the target to be
specified on the command line and that option will be recorded in the
commit message.

Test: ./update_prebuilts.py -p 7862434 --sdk_target sdk_phone_armv7-sdk
      ./update_prebuilts.py -f 32 7862434 --sdk_target sdk_phone_armv7-sdk
Change-Id: Ic01bfca2ac054a21a3a717f08f26e9d7677da7a8