Add 32 as the latest API directory
Adding a new API to a java_sdk_library will cause the library to try
and check that API against the previously released API to ensure
backwards compatibility. As the API is new there is no such previously
released API and so the build fails. The way to bypass that build
failure is to add an empty API specification in the latest release
directory.

Adding 32 as the latest API directory will make AOSP consistent with
downstream branches and allow new APIs to be added in a consistent way
in AOSP and downstream.

Bug: 216435117
Test: m checkapi
Change-Id: Ic504fad47f21fe124b73de0bd995088bf6cdb0e9