Delete manual BUILD files in current/extras and current/support.
in current/extras/app-toolkit/Android.bp, and
current/support/Android.bp, for each pair of library and import
(e.g. "android-arch-core-common-nodeps" and "android-arch-core-common"),
copy over static_libs of the library into the import. This is a no-op
in Soong because the imports are never depended on directly,
but it resolved an issue in Bazel where some of the imports are missing
needed direct dependencies. This is also the default behavior of pom2bp
so this shouldn't be altered by future updates (if any).

Test: b build //packages/apps/WallpaperPicker + CI
Change-Id: I43d77c1952924cd90b2c8290b3d0926db86a873d
Bug: 215230095
Bug: 215230093