Fixup jar files in extensions/1 dir
The art and conscrypt jar files were not named according to their
expected/new module names. Update the jar file names to match the txt
file names. Also remove the legacy/irrelevant platform APIs that are
not relevant to the public API.  It exists in many other places
(version control) for historical book-keeping if we need it.

Bug: 258836519
Test: presubmit
Change-Id: I3f056991d82ef6c5f78a531313d65f06c65f36be