Update finalized ART API to include bug fixes
A couple of bug fixes in metalava caused changes to the API txt file
for art which caused some Gantry gate failures because it detected a
change in the file and assumed it was an API change.

The simplest way to fix that is to just apply the same delta to the
already finalized API against which the gate is comparing the current
API. This change does just that, it modifies both the 34 and
extension 7 version of the files as they are identical and the patch
(see b/299080058 for the diff) applied 100% cleanly, i.e. no line
offsets.

FTR the ART changes being applied here are:
* https://r.android.com/2631169
* https://r.android.com/2594135.

Bug: 299080058
Test: leave it up to the gantry gate
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:791b5acc82f41eee4549eb443cb15281e8bc7bb9)
Merged-In: I178c73a209454392a65307f109265db9557f15b2
Change-Id: I178c73a209454392a65307f109265db9557f15b2