Update XmlResourceParser's last finalized api
Metalava behavior has changed w.r.t. inherited methods. Previously,
getAttributeNamespace was not present in txt files but it now is.

Replicate this change in the last finalized txt file, to avoid metalava
thinking this method has recently changed (the class has not been
modified since 2018).

Specifically, this fixes a problem where metalava otherwise thinks
getAttributeNamespace is a default method, which it is not because
it is being redefined in XmlResourceParser (as opposed to AttributeSet
where it is a default method).

Bug: 223382732
Test: m checkapi
Change-Id: I0c4f1e5172209ae71c5c7dac5bf415626b1b01e5