Update finalized ART API to include bug fixes
A bug fix in metalava caused a change to the API txt file for art which
caused a Gantry gate failure because it detected a change in the file
and assumed it was an API change.

The simplest way to fix that was to just apply the same delta to the
already finalized API against which the gate is comparing the current
API. This change does just that, it modifies both the 34 and
extension 7 version of the files as they are identical and the patch
(https://r.android.com/2973531) applied cleanly, i.e. with a minor line
offset.

Bug: 328236405
Bug: 326310840
Test: m checkapi
(cherry picked from https://android-review.googlesource.com/q/commit:55b81dcf11529e624f3318eb255e4fd5610b3c1c)
Merged-In: I819f4951372855e596da4924ee79a3f158cd0894
Change-Id: I819f4951372855e596da4924ee79a3f158cd0894