Only update android/os/ArtModuleServiceManager.class in platform SDK.
We need new APIs being in the prebuilt SDK so that we can build the ART
module on `master-art`.

Ideally, we should use the script to update the entire prebuilt SDK.
However, it's unfortunate that Metalava has a bug that it generates
wrong stubs for some classes, namely java.security.SecureRandom.
As a workaround, we have to manually update a single class since that's
all we need. Other classes are untouched.

The file is taken from
ab/11594605/sdk_with_runtime_apis-trunk_staging-userdebug

Change in readable format:

public class android.os.ArtModuleServiceManager {
  android.os.ArtModuleServiceManager();
  public android.os.ArtModuleServiceManager$ServiceRegisterer getArtdServiceRegisterer();
+ public android.os.ArtModuleServiceManager$ServiceRegisterer getArtdPreRebootServiceRegisterer();
+ public android.os.ArtModuleServiceManager$ServiceRegisterer getDexoptChrootSetupServiceRegisterer();
}

Bug: 311377497
Test: Presubmit
Test: (on master-art) -
  1. Patch https://r.android.com/3004314.
  2. Build the ART module.
Change-Id: Ia1b3d8d3031bd1ea0d2dd811d04a22a1da6b6250