Add android-plus-updatable.jar to module-lib and system-server
The `android.jar` in `module-lib` and `system-server` directories do
not include classes from the updatable modules. Also, while
`system-server` is supposed to be an extension of `module-lib` its
directories does not include system server APIs for the modules that
provide a `module-lib` API. e.g. `framework-permission.jar` exists
in `34/module-lib` but not in `34/system-server`. The latter only
includes `service-permission.jar`. That makes sense as only
`service-*.jar`s can provide a `system-server` API but the updatable
modules from `module-lib` are still part of the `system-server` API.

The lack of updatable classes in `android.jar` breaks the
`api-versions.xml` file generation for those API surfaces as that
requires a single jar containing the whole API for each surface or
each API version.

Unfortunately, it is not possible to modify the `android.jar` files
to include the updatable modules as other code depends on it not
including updatable modules. e.g. some modules will depend on a
specific prebuilt version of the `module-lib/android.jar`, e.g.
`sdk_version: "module_34"` and then depend on the source version of
an updatable jar, e.g. `framework-connectivity.stubs.module_lib` to
get the very latest version. Adding `framework-connectivity` to the
`android.jar` (which comes first on the classpath) would prevent the
compiler from seeing the latest version causing compilation errors.

This change adds a script and uses that script to add new
`android-plus-updatable.jar`s to the `module-lib` and `system-server`
directories as follows:

* The `module-lib/android-plus-updatable.jar` contains the contents of
  the `module-lib/android.jar` plus the contents of all the updatable
  module jars (excluding `art.jar` as that is already in `android.jar`
  otherwise it would not work as `javac` needs to see `java.lang.*`
  classes).
* The `system-server/android-plus-updatable.jar` contains the contents
  of the `system-server/android.jar` plus the contents of all the
  updatable module jars from `module-lib` plus the contents of all the
  `service-*.jar`s from `system-server`.

The script was run as follows:
    ./create-android-plus-updatable-jar.sh 30 31 32 33 34

The output can be found at https://paste.googleplex.com/6222208409534464.

Bug: 337836752
Test: TARGETS=$(for i in public system module-lib system-server; do U=$(echo $i | sed "s|-|_|g"); echo "out/target/common/obj/PACKAGING/api_versions_${U}_generated-api-versions.xml"; done)
      m $TARGETS; mkdir -p before; cp $TARGETS before
      m $TARGETS; mkdir -p after; cp $TARGETS after
      meld before after
      # Review differences.
Change-Id: I1eb55a974c555f7c527e461a5a19e1ef318955fa