Update baseline files for some ART/Bluetooth issues
The change in this topic corrects some issues with the compatibility
check of the combined API. Those issues were caused by the changes:
* https://r.android.com/2744916 - Added missing overrides to
  Bluetooth's `system-current.txt`
* https://r.android.com/2743812 - Added missing overrides to ART's
  `module-lib-current.txt`.

Previously, those changes were ignored by the
`frameworks-base-api-*-current-compat` targets but the other change in
this topic fixes that resulting in these issues being reported.

This change suppresses those issues by updating the baselines. This
should not be a problem once 35 is released and becomes the latest
version as that will include the missing overrides.

Bug: 320245343
Test: m checkapi
Change-Id: If300577d92a79a71e22ec0d1e07dc3526d1700ec