Make current_android_jar visible to aapt2 test app
In order to test some new features in aapt2, our build rules need to
call aapt2 manually in genrules. For this to work, we need to pass the
current android.jar to aapt2 but we need it visible to our project to do
that.

Test: Automated
Bug: 329436914
Flag: EXEMPT Aconfig not supported for visibility change
Change-Id: Ia395783c2a1ca15f80abcb5025adfe7f53efc793