Add missing type parameter to OutputConfiguration
Prior to 34 type parameters on constructors were not correctly recorded
in the signature files. Fortunately, only one constructor was affected,
the `OutputConfiguration`. This change fixes them through a combination
of manual changes and reapplying the resulting patch where possible.

The files were checked with the following:

```
  for i in */public/api/android.txt
  do
    metalava signature-cat $i > /dev/null
  done
```

It reported issues with version 26 (the version in which the affected
constructor was added) to 33 (inclusive - the last version affected by
the bug in Metalava).

Bug: 394597358
Test: # Ran the following after applying this change to each file to
      # make sure the error was fixed.
      metalava signature-cmd <version>/public/api/android.txt
Change-Id: I84895ad70c689617aaf17cc64fbfb2031574b75d