Inherit @Deprecated onto class members
Historically, the deprecated status that was written into signature
files was determined by what was written in the sources, which made the
signature files inconsistent. However, the deprecated status is now
inherited onto class members and nested classes making the signature
files more consistent. This change inherits deprecated status from a
deprecated class onto its class members.

Note: This does not inherit deprecated status from a class onto inner
classes as that is not easy to do using simple scripts. That will be
cleaned up in a follow-up change that uses `metalava signature-cat` to
reformat the files.

This extracts a change from the set of changes that reformatting using
`metalava signature-cat` would apply. It does that to make it easier to
review.

These changes were made using the following:
```
  # Find all the files that are affected.
  FILES=$(find -name android.txt -o -name android-removed.txt | xargs grep -l "^// Signature format:" | xargs grep -l " @Deprecated ")

  # Reformat the affected files and concatenate the result before making any changes.
  for i in $FILES; do /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat $i; done > before.txt 2> /dev/null

  # Apply the changes.
  sed -i "/  @Deprecated/,/}/s/\(    [^ ]\+\)/\1 @Deprecated/" $FILES

  # Remove any duplicate @Deprecated entries added by the previous command.
  sed -i "s/@Deprecated @Deprecated/@Deprecated/" $FILES

  # Reformat the affected files and concatenate the result after making these changes.
  # This ensures that the files are not broken.
  for i in $FILES; do /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat $i; done > after.txt 2> /dev/null

  # Make sure that these changes have not affected the final reformatted output.
  diff before.txt after.txt
```

Bug: 394597358
Test: See above.

Change-Id: Ib35302a69206dd199b025cf491db66bd8ac0b481