Remove redundant final keyword on methods in final classes
Historically, `final` keywords added in signature files were based on
their use in the source which made the signature files inconsistent in
their handling. However, that has since been fixed and the `final`
modifier is now normalized in signature files. Specifically, that means
that `final` modifiers are no longer written out on methods of `final`
classes. This change removes those `final` keywords.

This extracts a change from the set of changes that reformatting using
`metalava signature-cat` would apply. It does that to make it easier to
review.

These changes were made using the following:
```
  # Find all the files that are affected.
  FILES=$(find -name android.txt -o -name android-removed.txt | xargs grep -l "^// Signature format:" | xargs grep -l " final ")

  # Reformat the affected files and concatenate the result before making any changes.
  for i in $FILES; do /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat $i; done > before.txt 2> /dev/null

  # Apply the changes.
  sed -i "/^  [^ ].* final .* {/,/}/s/^\(    method .*\) final \(.*\)$/\1 \2/" $FILES

  # Reformat the affected files and concatenate the result after making these changes.
  # This ensures that the files are not broken.
  for i in $FILES; do /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat $i; done > after.txt 2> /dev/null

  # Make sure that these changes have not affected the final reformatted output.
  diff before.txt after.txt
```

Bug: 394597358
Test: See above.
Change-Id: Iae85022c3ca3879f42ffca03090fcb0e48c6fdba