Re-order the @Retention annotation before @Target
Historically, annotations were written to the signature files in the
order in which they were present in the source files which made the
signature files inconsistent. However, that has since been fixed and
annotations are now sorted before being written out. This only seems
to affect the order of the `@Target` and `@Retention` annotations
which this change fixes.

This extracts a change from the set of changes that reformatting using
`metalava signature-cat` would apply. It does that to make it easier to
review.

These changes were made using the following:
```
  # Find all the files that are affected.
  FILES=$(find -name android.txt -o -name android-removed.txt | xargs grep -l "^// Signature format:" | xargs grep -lE " @java\.lang\.annotation\.(Target|Retention)\(")

  # Reformat the affected files and concatenate the result before making any changes.
  for i in $FILES; do /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat $i; done > before.txt 2> /dev/null

  # Apply the changes.
  sed -i 's/\(@java.lang.annotation.Target([^)]*)\) \(@java.lang.annotation.Retention([^)]*)\)/\2 \1/' $FILES

  # Reformat the affected files and concatenate the result after making these changes.
  # This ensures that the files are not broken.
  for i in $FILES; do /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat $i; done > after.txt 2> /dev/null

  # Make sure that these changes have not affected the final reformatted output.
  diff before.txt after.txt
```

Bug: 394597358
Test: See above.
Change-Id: Ibaf30d3deda27e6454f7189a8526bb014cac6c3b