latest-finalized-flags: handle more than one .txt file
The genrule to calculate the most recent finalized-flags.txt file fails
if more than one finalized-flags.txt file exists because the file will
be listed on the same line. Add a call to `tr` to the pipe to convert
the space separated single line to multiple lines.

Also filter out current/finalized-flags.txt.

Bug: 377676163
Bug: 395581834
Test: m latest-finalized-flags
Change-Id: I600d198bdb75ed0fdde313a74b77fdc9e2b9751d