Remove extraneous space on package private classes
Historically, some package private classes did unintentionally leak
into the API signature files. The space that was intended to separate
the visibility modifier (`public` or `protected`) and the next
modifier or keyword was written out even when there was no visibility
modifier. That resulted in the class being indented 3 spaces instead of
the normal 2. The current code will only write the space separator out
if the visibility modifier is added.

This change fixes the indentation back to 2 spacing to reduce the size
of the follow-up change while will reformat the files using the current
code. Eventually, these classes will be removed altogether as they
should not be in the API but that cannot easily be done using a simple
script.

This extracts a change from the set of changes that reformatting using
`metalava signature-cat` would apply. It does that to make it easier to
review.

These changes were made using the following:
```
  # Find all the files that are affected.
  FILES=$(find -name android.txt -o -name android-removed.txt | xargs grep -l "^// Signature format:" | xargs grep -lE "^   [a-z]")

  # Reformat the affected files and concatenate the result before making any changes.
  for i in $FILES; do /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat $i; done > before.txt 2> /dev/null

  # Apply the changes.
  sed -i 's/^   \([a-z]\)/  \1/' $FILES

  # Reformat the affected files and concatenate the result after making these changes.
  # This ensures that the files are not broken.
  for i in $FILES; do /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat $i; done > after.txt 2> /dev/null

  # Make sure that these changes have not affected the final reformatted output.
  diff before.txt after.txt
```

Bug: 394597358
Test: See above.
Change-Id: I829cde9f09df54db7d025cd31cb8cae4853d296b