Reformat public android.txt files from 1 to 27
Reformats the public `android.txt` files from versions 1 to 27
(inclusive) using the same settings as is used in the build. The effect
of reformatting is to add `@Deprecated` annotations to inner classes
and their members that are not declared as deprecated but are inside a
class which is deprecated.

This is restricted to public APIs from version 1 to 27 because the
system APIs and versions after 28 use different formatting options.
They will be cleaned up in follow-up changes.

These changes were made using the following:
```
  # Find all the files that are affected.
  FILES=$(ls {1..27}/public/api/android.txt)

  # Apply the changes.
  for i in $FILES
  do
    /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat --format-defaults overloaded-method-order=source $i > tmp
    mv tmp $i
  done
```

Bug: 394597358
Test: See above.
Change-Id: Ie202326f524a0ccf359015588487ab8a20de109a