Reformat system android.txt files from 22 to 27
Reformats the system `android.txt` files from versions 22 (first) to 27
(inclusive) using some custom settings that differ from those used in
the build. The effect of reformatting is to add `@Deprecated`
annotations to inner classes and their members that are not declared as
deprecated but are inside a class which is deprecated.

This is restricted to system APIs from version 22 to 27 because the
system APIs after 28 use different formatting options. They will be
cleaned up in follow-up changes.

These changes were made using the following:
```
  # Find all the files that are affected.
  FILES=$(ls {22..27}/system/api/android.txt)

  # Apply the changes.
  for i in $FILES
  do
    /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat --format-defaults overloaded-method-order=source,strip-java-lang-prefix=never,type-argument-spacing=space $i > tmp 2>/dev/null
    mv tmp $i
  done
```

Bug: 394597358
Test: See above.
Change-Id: I699d3f57e31979a10e1aee47590f8ef2bd15526a