Add native keyword to invoke(...)/invokeExact(...)
The `MethodHandler.invoke(...)` and `MethodHandle.invokeExact(...)`
methods are treated specially by the compiler as long as they have the
`native` keyword. Generally, `native` was not considered part of the
API so it was not written out/removed, however in this case it is
significant so this change adds the `native` keyword.

See https://r.android.com/2807638 for more information.

This extracts a change from the set of changes that reformatting using
`metalava signature-cat` would apply. It does that to make it easier to
review.

These changes were made using the following:
```
  # Find all the files that are affected.
  FILES=$(find -name \*.txt | xargs grep -l "^// Signature format:" | xargs grep -lE "method public final Object invoke(Exact)?\(java\.lang\.Object\.\.\.\)")

  # Reformat the affected files and concatenate the result before making any changes.
  for i in $FILES; do /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat $i; done > before.txt 2> /dev/null

  # Apply the changes.
  perl -pi -e "s/(method public final) (Object invoke(Exact)?\(java\.lang\.Object\.\.\.\))/\1 native \2/" $FILES

  # Reformat the affected files and concatenate the result after making these changes.
  # This ensures that the files are not broken.
  for i in $FILES; do /etc/alternatives/java -jar /usr/local/google/home/paulduffin/android/aosp/main/out/host/linux-x86/framework/metalava.jar signature-cat $i; done > after.txt 2> /dev/null

  # Make sure that these changes have not affected the final reformatted output.
  diff before.txt after.txt
```

Bug: 394597358
Test: See above.
Change-Id: Ieece7f53dead91834be0d163bda8bb5f8db072bc