Generate 21/system/api/android.txt from android.jar
Previously, there was no signature file for the system API of 21 even
though that was the first release with a system API. This change
creates it from the `android.jar` using the following:

```
  mkdir 21/system/api/
  touch 21/system/api/android.txt
  /etc/alternatives/java -jar ../../out/host/linux-x86/framework/metalava.jar android-jars-to-signatures ../../ --format 2.0 --format-defaults overloaded-method-order=source,normalize-final-modifier=yes --config-file ../../build/soong/java/metalava/main-config.xml --api-versions 21 --api-surfaces system
```

It was tested by comparing with `22/system/api/android.txt` to see if
it was reasonably consistent with it. There were a few incompatible API
changes, e.g. fields being removed, method signatures being changed. To
be expected in the first version of the API. A sample of those
differences were checked in the source history and they appear to be
correct, e.g. `HdmiTimerRecordSources.Time` used to include `mHour` and
`mMinute` fields because they were `protected` fields of a `private`
`TimeUnit` super class. However, change ag/576254, made them package
private.

Bug: 394597358
Test: See above
Change-Id: I928bd8454fc20fab2f4af096b1ee9eebd468e648