Update -keepattributes *Annotation* rules in aars
This updates the rule `-keepattributes *Annotation*` in each of:

  runtime-28.0.0.aar
  design-28.0.0.aar
  coordinatorlayout-28.0.0.aar

... to:

-keepattributes AnnotationDefault,
                RuntimeVisibleAnnotations,
                RuntimeVisibleParameterAnnotations,
                RuntimeVisibleTypeAnnotations

The rule `-keepattributes *Annotation*` unintentionally matches RuntimeInvisibleAnnotations, RuntimeInvisibleParameterAnnotations and RuntimeInvisibleTypeAnnotations.

Such runtime invisible annotations cannot be looked up at runtime and thus should not be kept. In a future version of aosp it will be a hard error to keep runtime invisible annotations.

The proguard.txt file of each aar is updated manually (as opposed to updating these libraries to a newer version), since these legacy support libraries are no longer actively updated.

Bug: 387958004
Test: existing
Change-Id: I87afc3a870e2eaa1c0d2705cfd7cbfa79d018aa0