Reformat androidx-api.txt using metalava signature-cat
Previously, `androidx-api.txt` was created simply by concatenating all
the signature files from `frameworks/support` into a single file. That
is not strictly correct as the different files can have different
versions with different defaults which could make the result in
correct. Also, the order was not stable so every update would
essentially randomize the order.

This reformats it using the following command which will apply a stable
order and consistent formatting:
```
  metalava signature-cat --format 4.0 current/androidx-api.txt --output-file current/androidx-api.txt
```

A follow-up change will update the update script to use `signature-cat`
to correctly merge the files (and apply a stable order) and use that to
update the files with the latest version.

Bug: 402049952
Test: TH
Change-Id: Id4d60c104057a86c23fe4a03d0d797a6f05a95b2