update_prebuilts.py: remove dependency on distutils module
The update_prebuilts.py script relies on the distutils module, which was
removed in python 3.12. Attempting to run the script on machines running
later Python versions will cause this error:

  Traceback (most recent call last):
    File "/usr/local/google/home/amhk/android/main/./prebuilts/sdk/update_prebuilts.py", line 26, in <module>
      from distutils.version import LooseVersion
  ModuleNotFoundError: No module named 'distutils'

The parts of update_prebuilts.py that rely on distutils are related to
AndroidX drops, but AndroidX don't use this script anymore. Do the
simplest thing and remove the offending import statement: this would
cause the AndroidX code to crash if run, but no one is going to do that.

(And at some point the platform parts of update_prebuilts.py should be
rewritten in a more maintainable fashion, and at that point
update_prebuilts.py can be removed.)

Bug: 432423337
Test: croot && ./prebuilts/sdk/update_prebuilts.py # prints usage info, doesn't crash
Change-Id: I4dd541ac664a5fd1a6fa8d986478675600835197